+++ flow-server/src/main/java/com/vaadin/flow/server/DefaultDeploymentConfiguration.java
@@ -16,9 +16,7 @@
 
 package com.vaadin.flow.server;
 
-import java.util.Locale;
 import java.util.Properties;
-import java.util.function.Function;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,7 +32,7 @@ import com.vaadin.flow.shared.communication.PushMode;
  * @since 1.0
  */
 public class DefaultDeploymentConfiguration
-        extends AbstractDeploymentConfiguration {
+        extends PropertyDeploymentConfiguration {
     private static final String SEPARATOR = "\n===========================================================";
 
     public static final String NOT_PRODUCTION_MODE_INFO = SEPARATOR
@@ -73,17 +71,16 @@ public class DefaultDeploymentConfiguration
 
     public static final boolean DEFAULT_SEND_URLS_AS_PARAMETERS = true;
 
-    private final Properties initParameters;
     private boolean productionMode;
     private boolean xsrfProtectionEnabled;
     private int heartbeatInterval;
     private boolean closeIdleSessions;
     private PushMode pushMode;
     private String pushURL;
-    private final Class<?> systemPropertyBaseClass;
     private boolean syncIdCheck;
     private boolean sendUrlsAsParameters;
     private boolean requestTiming;
+    private static boolean warningsLogged = false;
 
     /**
      * Create a new deployment configuration instance.
@@ -97,8 +94,7 @@ public class DefaultDeploymentConfiguration
      */
     public DefaultDeploymentConfiguration(Class<?> systemPropertyBaseClass,
             Properties initParameters) {
-        this.initParameters = initParameters;
-        this.systemPropertyBaseClass = systemPropertyBaseClass;
+        super(systemPropertyBaseClass, initParameters);
 
         checkProductionMode();
         checkRequestTiming();
@@ -109,90 +105,7 @@ public class DefaultDeploymentConfiguration
         checkPushURL();
         checkSyncIdCheck();
         checkSendUrlsAsParameters();
-    }
-
-    @Override
-    public <T> T getApplicationOrSystemProperty(String propertyName,
-            T defaultValue, Function<String, T> converter) {
-        // Try system properties
-        String val = getSystemProperty(propertyName);
-        if (val != null) {
-            return converter.apply(val);
-        }
-
-        // Try application properties
-        val = getApplicationProperty(propertyName);
-        if (val != null) {
-            return converter.apply(val);
-        }
-
-        return defaultValue;
-    }
-
-    /**
-     * Gets an system property value.
-     *
-     * @param parameterName
-     *            the Name or the parameter.
-     * @return String value or null if not found
-     */
-    protected String getSystemProperty(String parameterName) {
-        String pkgName;
-        final Package pkg = systemPropertyBaseClass.getPackage();
-        if (pkg != null) {
-            pkgName = pkg.getName();
-        } else {
-            final String className = systemPropertyBaseClass.getName();
-            int index = className.lastIndexOf('.');
-            if (index >= 0) {
-                pkgName = className.substring(0, index);
-            } else {
-                pkgName = null;
-            }
-        }
-        if (pkgName == null) {
-            pkgName = "";
-        } else if (!pkgName.isEmpty()) {
-            pkgName += '.';
-        }
-        String val = System.getProperty(pkgName + parameterName);
-        if (val != null) {
-            return val;
-        }
-
-        // Try lowercased system properties
-        val = System.getProperty(
-                pkgName + parameterName.toLowerCase(Locale.ENGLISH));
-
-        if (val != null) {
-            return val;
-        }
-
-        // version prefixed with just "vaadin."
-        val = System.getProperty("vaadin." + parameterName);
-
-        return val;
-    }
-
-    /**
-     * Gets an application property value.
-     *
-     * @param parameterName
-     *            the Name or the parameter.
-     * @return String value or null if not found
-     */
-    public String getApplicationProperty(String parameterName) {
-
-        String val = initParameters.getProperty(parameterName);
-        if (val != null) {
-            return val;
-        }
-
-        // Try lower case application properties for backward compatibility with
-        // 3.0.2 and earlier
-        val = initParameters.getProperty(parameterName.toLowerCase());
-
-        return val;
+        warningsLogged = true;
     }
 
     /**
@@ -286,10 +199,6 @@ public class DefaultDeploymentConfiguration
         return pushURL;
     }
 
-    @Override
-    public Properties getInitParameters() {
-        return initParameters;
-    }
 
     /**
      * Log a warning if Vaadin is not running in production mode.
@@ -297,7 +206,7 @@ public class DefaultDeploymentConfiguration
     private void checkProductionMode() {
         productionMode = getBooleanProperty(
                 Constants.SERVLET_PARAMETER_PRODUCTION_MODE, false);
-        if (!productionMode) {
+        if (!productionMode && !warningsLogged) {
             getLogger().warn(NOT_PRODUCTION_MODE_INFO);
         }
     }
@@ -316,7 +225,7 @@ public class DefaultDeploymentConfiguration
     private void checkXsrfProtection() {
         xsrfProtectionEnabled = !getBooleanProperty(
                 Constants.SERVLET_PARAMETER_DISABLE_XSRF_PROTECTION, false);
-        if (!xsrfProtectionEnabled) {
+        if (!xsrfProtectionEnabled && !warningsLogged) {
             getLogger().warn(WARNING_XSRF_PROTECTION_DISABLED);
         }
     }
